+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+## TBD
+
+### Bug fixes
+
+* Set bugsnag.span.category to 'custom' for custom spans.
+  [336](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/336)
+
 ## 1.10.0 (2024-09-30)
 
 ### Enhancements